+++ test/Helper/Placeholder/ContainerTest.php
@@ -432,4 +432,20 @@ class ContainerTest extends TestCase
         $this->assertContains("    <ul>\n", $string, $string);
         $this->assertContains("\n    </ul>", $string, $string);
     }
+
+    /**
+     * @see https://github.com/zendframework/zend-view/pull/133
+     */
+    public function testNoPrefixOrPostfixAreRenderedIfNoItemsArePresentInTheContainer()
+    {
+        $this->container
+            ->setPrefix("<h1>")
+            ->setPostfix("</h1>");
+        $string = $this->container->toString();
+        $this->assertEquals('', $string);
+
+        $this->container->set('');
+        $string = $this->container->toString();
+        $this->assertEquals('<h1></h1>', $string);
+    }
 }